pyproject.toml
[project]
name = "freq-spectrum-viewer"
version = "0.1.0"
description = "UI application for image frequency spectrum analysis and comparison"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "streamlit>=1.28.0",
    "scipy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

